--- a/package.json
+++ b/package.json
@@ -10,7 +10,7 @@
         "build:dev": "rollup -c",
         "build": "NODE_ENV=production rollup -c",
         "watch": "npm start",
-        "test": "jest",
+        "test": "uvu -r esm ./tests",
         "cy:run": "cypress run",
         "cy:open": "cypress open",
         "format": "prettier './{packages/**/,cypress/**/*,}*.{js,html,css}' --write",
@@ -20,7 +20,6 @@
     },
     "license": "MIT",
     "devDependencies": {
-        "@babel/preset-env": "^7.9.5",
         "@rollup/plugin-node-resolve": "^7.1.3",
         "@rollup/plugin-replace": "^2.3.4",
         "@tailwindcss/forms": "^0.2.1",
@@ -29,8 +28,9 @@
         "cypress": "^6.6.0",
         "cypress-iframe": "^1.0.1",
         "edge.js": "^1.1.4",
+        "esm": "^3.2.25",
         "husky": "^4.3.0",
-        "jest": "^26.6.3",
+        "linkedom": "^0.5.6",
         "lint-staged": "^10.5.2",
         "prettier": "^2.2.0",
         "rollup": "^2.34.1",
@@ -39,7 +39,8 @@
         "rollup-plugin-postcss": "^4.0.0",
         "rollup-plugin-serve": "^1.1.0",
         "split-grid": "^1.0.9",
-        "tailwindcss": "^2.0.1"
+        "tailwindcss": "^2.0.1",
+        "uvu": "^0.5.1"
     },
     "dependencies": {},
     "husky": {