--- a/tests/utils.spec.js
+++ b/tests/utils.spec.js
@@ -1,45 +1,54 @@
+import { suite } from 'uvu'
+import * as assert from 'uvu/assert'
+import { parseHTML } from 'linkedom'
 import { isRequiredVersion, getComponentName } from '../packages/shell-chrome/src/utils'
 
-test('getComponentName > can handle multiple scenarios to determine component name', async () => {
+const ComponentNameTest = suite('getComponentName')
+
+ComponentNameTest('can handle multiple scenarios to determine component name', async () => {
+    const { document, window } = parseHTML(
+        `<div x-title="foo" x-id="bar" id="baz" name="qux" aria-label="quux" x-data="myFn(param)" role="quuz"></div>`,
+    )
     window.myFn = () => {}
-    document.body.innerHTML = `
-        <div x-title="foo" x-id="bar" id="baz" name="qux" aria-label="quux" x-data="myFn(param)" role="quuz"></div>
-    `
 
     const element = document.querySelector('[x-data]')
 
-    expect(getComponentName(element)).toBe('foo')
+    assert.equal(getComponentName(element), 'foo')
     element.removeAttribute('x-title')
-    expect(getComponentName(element)).toBe('bar')
+    assert.equal(getComponentName(element), 'bar')
     element.removeAttribute('x-id')
-    expect(getComponentName(element)).toBe('baz')
+    assert.equal(getComponentName(element), 'baz')
     element.removeAttribute('id')
-    expect(getComponentName(element)).toBe('qux')
+    assert.equal(getComponentName(element), 'qux')
     element.removeAttribute('name')
-    expect(getComponentName(element)).toBe('quux')
+    assert.equal(getComponentName(element), 'quux')
     element.removeAttribute('aria-label')
-    expect(getComponentName(element)).toBe('myFn')
+    assert.equal(getComponentName(element), 'myFn')
     element.setAttribute('x-data', '{ foo: "bar" }')
-    expect(getComponentName(element)).toBe('quuz')
+    assert.equal(getComponentName(element), 'quuz')
     element.removeAttribute('role')
-    expect(getComponentName(element)).toBe('div')
+    assert.equal(getComponentName(element), 'div')
 })
 
-describe('isRequiredVersion', () => {
-    test('works for major', () => {
-        expect(isRequiredVersion('1.11.0', '2.1.1')).toBe(true)
-        expect(isRequiredVersion('2.11.0', '0.1.1')).toBe(false)
-    })
-    test('works for minor', () => {
-        expect(isRequiredVersion('1.1.1', '1.11.0')).toBe(true)
-        expect(isRequiredVersion('1.11.0', '1.1.1')).toBe(false)
-    })
-    test('works for patch', () => {
-        expect(isRequiredVersion('0.11.0', '0.11.1')).toBe(true)
-        expect(isRequiredVersion('0.1.1', '0.1.0')).toBe(false)
-    })
-    test('works for equal', () => {
-        expect(isRequiredVersion('2.8.0', '2.8.0')).toBe(true)
-        expect(isRequiredVersion('1.1.0', '1.1.0')).toBe(true)
-    })
+ComponentNameTest.run()
+
+const RequiredVersionTest = suite('isRequiredVersion')
+
+RequiredVersionTest('works for major', () => {
+    assert.equal(isRequiredVersion('1.11.0', '2.1.1'), true)
+    assert.equal(isRequiredVersion('2.11.0', '0.1.1'), false)
 })
+RequiredVersionTest('works for minor', () => {
+    assert.equal(isRequiredVersion('1.1.1', '1.11.0'), true)
+    assert.equal(isRequiredVersion('1.11.0', '1.1.1'), false)
+})
+RequiredVersionTest('works for patch', () => {
+    assert.equal(isRequiredVersion('0.11.0', '0.11.1'), true)
+    assert.equal(isRequiredVersion('0.1.1', '0.1.0'), false)
+})
+RequiredVersionTest('works for equal', () => {
+    assert.equal(isRequiredVersion('2.8.0', '2.8.0'), true)
+    assert.equal(isRequiredVersion('1.1.0', '1.1.0'), true)
+})
+
+RequiredVersionTest.run()