uvu + esm + linkedom for HTML/DOM

- test runs are faster (although it's splitting hairs, Jest was reporting 2s, uvu <1s)
- we don't use all the Jest features
- 600 fewer packages to install probably due to Babel/JSDOM more than Jest
  - before `added 1448 packages in 14.321s`
  - after `added 828 packages in 11.358s`

This change is up for debate Jest + Babel + JSDOM is probably the more widely adopted so happy not to make this switch.

Our key tests are written using Cypress so we could drop this suite altogether.