@HugoDF 2021-03-19
> Looks fine to me. The syntax is basically the same? I haven't heard of uvu but I'm always happy to try out new things

it's a bit more explicit, eg. "import {suite} from 'uvu'" opposed to Jest injecting globals

doesn't seem like we're adding many unit tests in any case 😄